+++ CHANGELOG.rst
@@ -21,6 +21,7 @@ master
 Changed
 ~~~~~~~
 
+- (`#46 <https://github.com/openscm/openscm-runner/pull/46>`_) Use `pytest markers <https://docs.pytest.org/en/stable/example/markers.html>`_ for marking tests which rely on MAGICC rather than hack fixture solution
 - (`#45 <https://github.com/openscm/openscm-runner/pull/45>`_) Update regression tests so they can be more easily updated
 - (`#44 <https://github.com/openscm/openscm-runner/pull/44>`_) Updated tests to using MAGICCv7.5.1